+++ clickcoin/src/main/java/clicker/observer/ObserverEnemigo.java
@@ -10,7 +10,7 @@ package clicker.observer;
  * @author Nehemias
  */
 public interface ObserverEnemigo {
-    public void quitarBTC(int cantidad);
-    public void quitarETH(int cantidad);
+    public void quitarBTC(float cantidad);
+    public void quitarETH(float cantidad);
     
 }
